requirements.txt
torch~=2.4.1+cu121
torchvision~=0.19.1
fastapi
uvicorn
pillow~=10.4.0
flask~=3.0.3